+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change log for [LogESP](https://github.com/dogoncouch/LogESP)
 ## [Unreleased]
 ### Fixed
 - `make update` reboot option
+- Recurring parser DB errors
 
 ## [0.2-beta] 2018-04-04
 ### Added